requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.0
pytz==2018.3
